+++ packages/asgardex-binance/lib/index.d.ts
@@ -1,4 +1,0 @@
-export * from './client';
-export * from './util';
-export * from './types/binance';
-export * as WS from './types/binance-ws';