+++ packages/asgardex-binance/src/types/binance.ts
@@ -385,6 +385,11 @@ export type GetTxsParams = {
   txType?: TxType
 }
 
+export type VaultTxParams = { addressFrom?: Address; addressTo: Address; amount: number; asset: string; memo: string }
+export type NormalTxParams = { addressFrom?: Address; addressTo: Address; amount: number; asset: string }
+export type FreezeParams = { address?: Address; amount: number; asset: string }
+export type MultiSendParams = { address: Address; transactions: MultiTransfer[]; memo?: string }
+export type GetMarketsParams = { limit?: number; offset?: number }
 /**
  * OrderList
  * @see https://docs.binance.org/api-reference/dex-api/paths.html#orderlist