+++ packages/asgardex-binance/CHANGELOG.md
@@ -1,3 +1,24 @@
+# v.3.0.0 (2020-08-26)
+
+### Breaking changes:
+- `Constructor` argument is an object now `{ network: Network; phrase?: string }`
+- `getAddress` returns undefined if `phrase` has not been set before
+- `getPrivateKey()` throws an error if `phrase` has not been set before
+- `setPrivateKey` rejects if a `phase` has not been set before
+- `vaultTx`, `normalTx`, `multiSend`, `getMarkets` accept an object as its parameters
+
+### Add:
+- `freeze(params: FreezeParams): Promise<TransferResult>`
+- `unfreeze(params: FreezeParams): Promise<TransferResult>`
+- `getBncClient(): BncClient`
+
+### Fix:
+- Fix `Rollup` warnings of `Unresolved dependencies` and `Circular dependencies`
+
+### Update:
+- Use latest npm dependencies
+
+
 # v.2.1.1 (2020-08-14)
 
 - Fix result type of `getFees()`